Update cylinder flow SVD tutorial to SmartSim 0.8.0.
The latested SmartSim version does not allow for duplicated model
names within a single Experiment. Moreover, each model creates a
new subfolder corresponding to the model name when the model is
executed. The subfolder creation can be avoided by passing a the
*path* agrument to Experiment.create_model(). Naming conflicts
arising when computing the SVD with different values for the rank
truncation parameter are now avoided by including the rank parameter
in the model name.